requirements.txt
absl-py @ file:///croot/absl-py_1666362940888/work
aiohttp @ file:///tmp/build/80754af9/aiohttp_1646806365504/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
anyio==3.6.2
astor==0.8.1
astunparse==1.6.3
async-timeout @ file:///opt/conda/conda-bld/async-timeout_1664876359750/work
attrs @ file:///croot/attrs_1668696182826/work
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///tmp/build/80754af9/click_1646056590078/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
conda==23.1.0
conda-content-trust @ file:///tmp/build/80754af9/conda-content-trust_1617045594566/work
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1649105784853/work
cryptography @ file:///tmp/build/80754af9/cryptography_1639414572950/work
enum34==1.1.10
fastapi==0.94.1
flatbuffers @ file:///home/ktietz/cip/python-flatbuffers_1634039120618/work
frozenlist @ file:///tmp/build/80754af9/frozenlist_1637767113340/work
gast @ file:///Users/ktietz/demo/mc3/conda-bld/gast_1628588903283/work
google-auth @ file:///opt/conda/conda-bld/google-auth_1646735974934/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1617120569401/work
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
grpcio @ file:///tmp/build/80754af9/grpcio_1637590821884/work
h11==0.14.0
h5py @ file:///tmp/build/80754af9/h5py_1608649014695/work
httptools==0.5.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1648544546694/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
Markdown @ file:///croot/markdown_1671541909495/work
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
multidict @ file:///opt/conda/conda-bld/multidict_1640703752579/work
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1651563629415/work
oauthlib @ file:///croot/oauthlib_1665490897403/work
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
pluggy @ file:///tmp/build/80754af9/pluggy_1648024445381/work
protobuf==3.20.1
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.6
PyJWT @ file:///opt/conda/conda-bld/pyjwt_1657544592787/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-arango==7.5.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.0
requests-toolbelt==0.10.1
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016711199/work
ruamel.yaml @ file:///tmp/build/80754af9/ruamel.yaml_1613650277144/work
ruamel.yaml.clib @ file:///opt/conda/conda-bld/ruamel.yaml.clib_1641552548372/work
scipy @ file:///tmp/build/80754af9/scipy_1641555004408/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
starlette==0.26.1
tensorboard @ file:///home/builder/stiwari/miniconda3/envs/tf210/conda-bld/tensorboard_1666889856225/work/tensorboard-2.10.0-py3-none-any.whl
tensorboard-data-server @ file:///tmp/build/80754af9/tensorboard-data-server_1633034767313/work/tensorboard_data_server-0.6.0-py3-none-manylinux2010_x86_64.whl
tensorboard-plugin-wit @ file:///home/builder/tkoch/workspace/tensorflow/tensorboard-plugin-wit_1658918494740/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorflow @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_cpu/tensorflow-base_1614360980945/work/tensorflow_pkg/tensorflow-2.4.1-cp39-cp39-linux_x86_64.whl
tensorflow-estimator @ file:///home/builder/adipietro/tf/tensorflow-estimator_1630508970172/work/tensorflow_estimator-2.6.0-py2.py3-none-any.whl
termcolor @ file:///croot/termcolor_1668084651543/work
toolz @ file:///croot/toolz_1667464077321/work
tqdm @ file:///opt/conda/conda-bld/tqdm_1647339053476/work
typing_extensions==4.5.0
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0
Werkzeug @ file:///croot/werkzeug_1671215996852/work
wrapt @ file:///tmp/build/80754af9/wrapt_1638433857881/work
yarl @ file:///tmp/build/80754af9/yarl_1606939947528/work
zipp @ file:///croot/zipp_1672387121353/work
